Remove unnecessary dependencies and config
There's no code in this repository, so it doesn't make sense to declare dependencies on large libraries like pandas and cartopy if there's nothing in the repo using them.
Dependencies should be declared alongside the code that uses them (presumably in clean-air-project. I imagine this is just a copy & paste thing rather than intentional.